refactor(types): rename t -> typings
