+++ src/routers/organizations.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends
 
 from db import methods
-from t import types, RQ, RS
+from typings import types, RQ, RS
 from utils.auth.auth import get_current_user
 from utils.response import ErrorCodes, ErrorResponse, SuccessfulResponse
