+++ src/utils/auth/auth.py
@@ -5,7 +5,7 @@ from passlib.context import CryptContext
 from fastapi import Header
 from fastapi import status as http_status
 
-from t import types
+from typings import types
 from db import methods
 from utils import response
 from config import config
